feat(example): 新增 mica-ai 全能力集成示例与自动装配改进
- 新增 mica-ai-example 模块，包含 Spring Boot 集成测试及 Demo 启动入口
- example 模块集中引入 mica-ai 全部 6 个 Starter 依赖，提供可运行示例
- 配置示例 application.yml，支持按需启用各模型及路径配置
- 实现 FaceController 和 IntentController 提供人脸识别与意图分类 REST 接口
- 优化各 Starter 自动配置，新增 enabled 开关支持，缺少必填项时 fail-fast 抛出异常
- 增加 OpenCVNativeLoader 组件，确保 mica-ai-ppocr 依赖的 OpenCV native 库提前加载
- 修正 mica-ai-ppocr DbPostProcessor 中的 shape 类型转换问题
- 新增 ExampleApplicationContextTest，测试各 Starter 启用禁用及配置异常场景
- 更新根 pom.xml，支持 example 模块与全 Starter 依赖管理
- 新增 Logback 配置文件，支持日志文件滚动和日志级别管理

+++ mica-ai-core/mica-ai-ppocr/src/main/java/net/dreamlu/mica/ai/ppocr/postprocessor/DbPostProcessor.java
@@ -37,8 +37,8 @@ public final class DbPostProcessor {
 		int srcW = (int) imgShape[1];
 
 		if (prob.dims() == 4 && prob.size(0) == 1 && prob.size(1) == 1) {
-			int hNew = (int) prob.size(2);
-			int wNew = (int) prob.size(3);
+			int hNew = prob.size(2);
+			int wNew = prob.size(3);
 			prob = prob.reshape(1, hNew);
 			if (prob.cols() != wNew) {
 				prob = prob.reshape(1, wNew);
